Make TemplateRegistry not mutable
Update UPGRADE-3.x.md
