+++ src/Templating/MutableTemplateRegistryInterface.php
@@ -19,15 +19,26 @@ namespace Sonata\AdminBundle\Templating;
 interface MutableTemplateRegistryInterface extends TemplateRegistryInterface
 {
     /**
+     * NEXT_MAJOR: remove this method declaration with docblock and uncomment code below.
+     *
      * @param array<string, string> $templates 'name' => 'file_path.html.twig'
      */
     public function setTemplates(array $templates);
 
+    ///**
+    // * @param array<string, string> $templates 'name' => 'file_path.html.twig'
+    // */
+    //public function setTemplates(array $templates): void;
+
     /**
+     * NEXT_MAJOR: remove this method declaration with docblock and uncomment code below.
+     *
      * @param string $name
      * @param string $template
      *
      * @return void
      */
     public function setTemplate($name, $template);
+
+    //public function setTemplate(string $name, string $template): void;
 }
